refactor policies to model, create reusable bindings
