+++ libs/model/src/policies/index.ts
@@ -1,3 +1,6 @@
 export * from './ContestWorker.policy';
 export * from './DiscordBot.policy';
 export * from './FarcasterWorker.policy';
+export * from './chainEventCreatedPolicy';
+export * from './handleCommunityStakeTrades';
+export * from './handleLaunchpadTrade';
