+++ packages/commonwealth/server/workers/commonwealthConsumer/commonwealthConsumer.ts
@@ -1,37 +1,17 @@
 import {
   HotShotsStats,
-  RabbitMQAdapter,
-  RascalConfigServices,
   ServiceKey,
-  buildRetryStrategy,
-  getRabbitMQConfig,
   startHealthCheckLoop,
 } from '@hicommonwealth/adapters';
-import {
-  Broker,
-  BrokerSubscriptions,
-  broker,
-  handleEvent,
-  logger,
-  stats,
-} from '@hicommonwealth/core';
-import {
-  Contest,
-  ContestWorker,
-  DiscordBotPolicy,
-  FarcasterWorker,
-  User,
-} from '@hicommonwealth/model';
+import { handleEvent, logger, stats } from '@hicommonwealth/core';
+import { ContestWorker } from '@hicommonwealth/model';
 import { EventNames } from '@hicommonwealth/schemas';
+import { bootstrapBindings } from 'server/bindings/bootstrap';
 import { fileURLToPath } from 'url';
-import { config } from '../../config';
-import { ChainEventPolicy } from './policies/chainEventCreated/chainEventCreatedPolicy';
 
 const log = logger(import.meta);
 
-stats({
-  adapter: HotShotsStats(),
-});
+stats({ adapter: HotShotsStats() });
 
 let isServiceHealthy = false;
 
@@ -45,17 +25,6 @@ startHealthCheckLoop({
   },
 });
 
-function checkSubscriptionResponse(
-  subRes: boolean,
-  topic: BrokerSubscriptions,
-) {
-  if (!subRes) {
-    log.fatal(`Failed to subscribe to ${topic}. Requires restart!`, undefined, {
-      topic,
-    });
-  }
-}
-
 // CommonwealthConsumer is a server that consumes (and processes) RabbitMQ messages
 // from external apps or services (like the Snapshot Service). It exists because we
 // don't want to modify the Commonwealth database directly from external apps/services.
@@ -63,91 +32,6 @@ function checkSubscriptionResponse(
 // to the CommonwealthConsumer and you want to ensure that the CommonwealthConsumer is
 // properly handling/processing those messages. Using the script is rarely necessary in
 // local development.
-
-export async function setupCommonwealthConsumer(): Promise<void> {
-  let brokerInstance: Broker;
-  try {
-    const rmqAdapter = new RabbitMQAdapter(
-      getRabbitMQConfig(
-        config.BROKER.RABBITMQ_URI,
-        RascalConfigServices.CommonwealthService,
-      ),
-    );
-    await rmqAdapter.init();
-    broker({
-      adapter: rmqAdapter,
-    });
-    brokerInstance = rmqAdapter;
-  } catch (e) {
-    log.error(
-      'Rascal consumer setup failed. Please check the Rascal configuration',
-    );
-    throw e;
-  }
-
-  const chainEventSubRes = await brokerInstance.subscribe(
-    BrokerSubscriptions.ChainEvent,
-    ChainEventPolicy(),
-  );
-  checkSubscriptionResponse(chainEventSubRes, BrokerSubscriptions.ChainEvent);
-
-  const contestWorkerSubRes = await brokerInstance.subscribe(
-    BrokerSubscriptions.ContestWorkerPolicy,
-    ContestWorker(),
-    buildRetryStrategy(undefined, 20_000),
-    {
-      beforeHandleEvent: (topic, event, context) => {
-        context.start = Date.now();
-      },
-      afterHandleEvent: (topic, event, context) => {
-        const duration = Date.now() - context.start;
-        const handler = `${topic}.${event.name}`;
-        stats().histogram(`cw.handlerExecutionTime`, duration, { handler });
-      },
-    },
-  );
-  checkSubscriptionResponse(
-    contestWorkerSubRes,
-    BrokerSubscriptions.ContestWorkerPolicy,
-  );
-
-  const contestProjectionsSubRes = await brokerInstance.subscribe(
-    BrokerSubscriptions.ContestProjection,
-    Contest.Contests(),
-  );
-  checkSubscriptionResponse(
-    contestProjectionsSubRes,
-    BrokerSubscriptions.ContestProjection,
-  );
-
-  const xpProjectionSubRes = await brokerInstance.subscribe(
-    BrokerSubscriptions.XpProjection,
-    User.Xp(),
-  );
-  checkSubscriptionResponse(
-    xpProjectionSubRes,
-    BrokerSubscriptions.XpProjection,
-  );
-
-  const farcasterWorkerSubRes = await brokerInstance.subscribe(
-    BrokerSubscriptions.FarcasterWorkerPolicy,
-    FarcasterWorker(),
-    buildRetryStrategy(undefined, 20_000),
-  );
-  checkSubscriptionResponse(
-    farcasterWorkerSubRes,
-    BrokerSubscriptions.FarcasterWorkerPolicy,
-  );
-
-  const discordBotSubRes = await brokerInstance.subscribe(
-    BrokerSubscriptions.DiscordBotPolicy,
-    DiscordBotPolicy(),
-  );
-  checkSubscriptionResponse(
-    discordBotSubRes,
-    BrokerSubscriptions.DiscordBotPolicy,
-  );
-}
 
 function startRolloverLoop() {
   log.info('Starting rollover loop');
@@ -172,7 +56,7 @@ function startRolloverLoop() {
 async function main() {
   try {
     log.info('Starting main consumer');
-    await setupCommonwealthConsumer();
+    await bootstrapBindings();
     isServiceHealthy = true;
     startRolloverLoop();
   } catch (error) {
