+++ packages/commonwealth/test/integration/commonwealthConsumer/chainEventCreatedPolicy.spec.ts
@@ -1,11 +1,9 @@
 import { EventContext, dispose } from '@hicommonwealth/core';
-import { DB, tester } from '@hicommonwealth/model';
+import { DB, processChainEventCreated, tester } from '@hicommonwealth/model';
 import { BalanceType } from '@hicommonwealth/shared';
 import { expect } from 'chai';
 import { BigNumber } from 'ethers';
 import { afterAll, afterEach, beforeAll, describe, test } from 'vitest';
-// eslint-disable-next-line max-len
-import { processChainEventCreated } from '../../../server/workers/commonwealthConsumer/policies/chainEventCreated/chainEventCreatedPolicy';
 
 // These are all values for a real txn on the Ethereum Sepolia Testnet
 const transactionHash =
